다른 유저 수정, 삭제 막기

+++ src/controllers/post.controller.js
@@ -4,15 +4,23 @@ const {
   getPost,
   updatePost,
   deletePost,
-  getBoardId,
+  matchOwner,
   getUserPostsCount,
   getPostsByUserId,
 } = require('../services/post.service');
+const { getBoardId } = require('../services/board.service');
 const postController = require('express').Router();
 
 postController.post('/create', withAuth, async (req, res) => {
   const { url, title, content, type } = req.body;
   const creator = req.userId;
+
+  if (!creator) {
+    return res.status(401).json({
+      isError: true,
+      message: '로그인 유지시간이 만료되었습니다. 다시 로그인해주세요.',
+    });
+  }
 
   if (!title || !content) {
     return res
@@ -31,7 +39,7 @@ postController.post('/create', withAuth, async (req, res) => {
     const post = await createPost({ creator, boardId, title, content, type });
 
     if (post.errorMsg !== null) {
-      throw new Error(post.errorMsg);
+      res.status(500).json({ isError: true, message: isOwnerData.errorMsg });
     }
 
     return res
@@ -45,11 +53,19 @@ postController.post('/create', withAuth, async (req, res) => {
 
 postController.get('/detail/:postId', withAuth, async (req, res) => {
   const { postId } = req.params;
+  const creator = req.userId;
 
   if (!postId) {
     return res
       .status(400)
       .json({ isError: true, message: '등록된 게시글이 없습니다.' });
+  }
+
+  if (!creator) {
+    return res.status(401).json({
+      isError: true,
+      message: '로그인 유지시간이 만료되었습니다. 다시 로그인해주세요.',
+    });
   }
 
   try {
@@ -62,6 +78,38 @@ postController.get('/detail/:postId', withAuth, async (req, res) => {
     return res.status(200).json({ post: postData.post });
   } catch (err) {
     console.error(`[post/detail]: ${err}`);
+    return res.status(500).json({ isError: true, message: `${err}` });
+  }
+});
+
+postController.get('/checkuser/:postId', withAuth, async (req, res) => {
+  const { postId } = req.params;
+  const creator = req.userId;
+  if (!creator) {
+    return res.status(401).json({
+      isError: true,
+      message: '로그인 유지시간이 만료되었습니다. 다시 로그인해주세요.',
+    });
+  }
+
+  if (!postId) {
+    return res
+      .status(400)
+      .json({ isError: true, message: '게시글 정보가 없습니다.' });
+  }
+
+  try {
+    const isOwnerData = await matchOwner({ postId, creator });
+
+    if (isOwnerData.errorMsg !== null) {
+      res.status(500).json({ isError: true, message: isOwnerData.errorMsg });
+    }
+
+    if (isOwnerData.isOwner) {
+      return res.status(200).json({ isError: false, message: '작성자입나다.' });
+    }
+  } catch (err) {
+    console.error(`[post/checkuser/:postId]: ${err}`);
     return res.status(500).json({ isError: true, message: `${err}` });
   }
 });
@@ -90,15 +138,23 @@ postController.put('/update/:postId', withAuth, async (req, res) => {
   }
 
   try {
-    const post = await updatePost({ postId, title, content });
+    const isOwnerData = await matchOwner({ postId, creator });
 
-    if (post.errorMsg !== null) {
-      throw new Error(post.errorMsg);
+    if (isOwnerData.errorMsg !== null) {
+      res.status(500).json({ isError: true, message: isOwnerData.errorMsg });
     }
 
-    return res
-      .status(200)
-      .json({ isError: false, message: '게시글이 수정되었습니다.' });
+    if (isOwnerData.isOwner) {
+      const post = await updatePost({ postId, title, content });
+
+      if (post.errorMsg !== null) {
+        throw new Error(post.errorMsg);
+      }
+
+      return res
+        .status(200)
+        .json({ isError: false, message: '게시글이 수정되었습니다.' });
+    }
   } catch (err) {
     console.error(`[post/update]: ${err}`);
     return res.status(500).json({ isError: true, message: `${err}` });
@@ -122,15 +178,23 @@ postController.delete('/:postId', withAuth, async (req, res) => {
   }
 
   try {
-    const postData = await deletePost({ postId });
+    const isOwnerData = await matchOwner({ postId, creator });
 
-    if (postData.errorMsg !== null) {
-      throw new Error(post.errorMsg);
+    if (isOwnerData.errorMsg !== null) {
+      res.status(500).json({ isError: true, message: isOwnerData.errorMsg });
     }
 
-    return res
-      .status(200)
-      .json({ isError: false, message: '게시글이 삭제되었습니다.' });
+    if (isOwnerData.isOwner) {
+      const postData = await deletePost({ postId });
+
+      if (postData.errorMsg !== null) {
+        throw new Error(post.errorMsg);
+      }
+
+      return res
+        .status(200)
+        .json({ isError: false, message: '게시글이 삭제되었습니다.' });
+    }
   } catch (err) {
     console.error(`[post/:postId]: ${err}`);
     return res.status(500).json({ isError: true, message: `${err}` });
